Move appState to common module, make case for object consistent
